feat: refactor FlashCardPage to use FlashCardClient for improved UX and loading state
